Add header file with link to upper level. Use TAGFILES based on DEPENDS argument to create_doxygen_documentation() macro.
Notes:
create_doxygen_documentation(DEPENDS artdaq)
The DEPENDS argument should be a list of packages which the current
package depends upon. This information will be used to link the package
documentation together.

The basic assumption being made is that each package has its html
documentation in a directory named after the package, all at the same
level, and that an index.html also exists at that level. For example:

+-- docs
 |- index.html
 |- artdaq/
  |- index.html
 |- artdaq_utilities/
  |- index.html
...